requirements-42.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
Django==4.2.10
django-appconf==1.0.6
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==23.5
django-models-logging==2.3.2
django-o365mail==1.1.0
idna==3.6
numpy==1.26.4
O365==2.0.32
oauthlib==3.2.2
pandas==2.2.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stringcase==1.2.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
